requirements.txt
fastapi>=0.109.0
pydantic>=2.0.0
uvicorn>=0.27.0
anthropic>=0.40.0
python-docx>=1.1.0
ebooklib>=0.18
lxml>=4.9.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
Pillow>=10.0.0
